VGG16 Compilation, convertImages.jl improvements
VGG16 model has been implemented in Knet. vgg16.jl now compiles the
model and saves it into a .jld file (~500MB). convertImages.jl has been
renamed into videotranslation.jl as it will be the main script. The code
has been modified to be ready for the addition of the VGG model, and the
LSTM model (will be implemented in a later step)